Update __init__.py
modify some bugs/features:
1. original: if there's a quote tweet in a reply, then toot it.
change to: ignore replies anyway.

2. original: if there's a quote, then ignore the photos outside the quote.
change to: attach the photos from the quote tweet behind the media list of the new tweet. 

3. original: do not parse videos, parse thumbnail instead.
change to: upload mp4 video in the toot. If a video is found, then ignore all the other photos.
some mp4 got from Twitter are encoded as iso5 format, which is not accepted by Mastodon. So, if the upload failed, convert the video to mp42 format with local FFmpeg, then reupload.